fix: 修复了传入字符串无法识别写入牌堆的问题

+++ task_control_execute.h
@@ -1,6 +1,5 @@
 #ifndef TASK_CONTROL_EXECUTE_H
 #define TASK_CONTROL_EXECUTE_H
-
 #include <algorithm>
 #include <iostream>
 #include <string>
@@ -9,14 +8,12 @@
 #include "calculator.h"
 
 #define MAX_CARD_SUIT_NUM (8)
-
-enum class SwitchCardType : uint8_t {
+enum class SwitchCardType : int {
     kUnknowType = 0,
     kSwitchByIndex = 1,
     kSwitchById = 2
 };
-
-enum class SubCmdType : uint8_t {
+enum class SubCmdType : int {
     kNoneType = 0,
     kAddCards = 1,
     kDeleteCards = 2,
@@ -24,15 +21,13 @@ enum class SubCmdType : uint8_t {
     kLoadCards = 4,
     kResetAllCards = 5
 };
-
 typedef struct {
-    uint8_t subCmd;
-    uint8_t switch_type;
-    uint8_t cards_index;
+    int subCmd;
+    int switch_type;
+    int cards_index;
     std::string cards_id;
-    std::vector<uint8_t> operate_cards;
+    std::vector<int> operate_cards;
 } cards_receive_data_t;
-
 class TaskControlExecute {
    public:
     TaskControlExecute();
@@ -47,24 +42,23 @@ class TaskControlExecute {
     void update(const cards_receive_data_t &data);
 
    private:
-    bool checkIfIndexValid(uint8_t index);
+    bool checkIfIndexValid(int index);
     bool checkIfIdValid(std::string id);
-    void add(uint8_t index, std::string id, const std::vector<uint8_t> &cards);
-    void del(uint8_t index, const std::vector<uint8_t> &cards);
-    void del(std::string id, const std::vector<uint8_t> &cards);
-    void modify(uint8_t index, const std::vector<uint8_t> &cards);
-    void modify(std::string id, const std::vector<uint8_t> &cards);
-    CardsEventCalculator load(uint8_t index);
+    void add(int index, std::string id, const std::vector<int> &cards);
+    void del(int index, const std::vector<int> &cards);
+    void del(std::string id, const std::vector<int> &cards);
+    void modify(int index, const std::vector<int> &cards);
+    void modify(std::string id, const std::vector<int> &cards);
+    CardsEventCalculator load(int index);
     CardsEventCalculator load(std::string id);
-    void addCalculator(uint8_t index);
+    void addCalculator(int index);
     void addCalculator(std::string id);
-    void addCalculator(uint8_t index, std::string id);
+    void addCalculator(int index, std::string id);
     void resetAllCalculator() { cards_calculator_.clear(); };
-    double getCalculatorPairProbability(uint8_t index);
+    double getCalculatorPairProbability(int index);
     double getCalculatorPairProbability(std::string id);
 
    private:
     std::vector<CardsEventCalculator> cards_calculator_;
 };
-
 #endif  // TASK_CONTROL_EXECUTE_H
